further attachment refactor
